Commenting out an old test in RawMessageCodecTest (#18)
It seems that the signature is no longer verifying when extra bytes are added to the end after changing the crypto library. In the past, this verified because the length is listed in the signature and it only checked that many bytes. Now it throws an exception for a parsing error, which is reasonable behavior. Because this tests the crypto library behavior and not our code, and because it is incosistent across different libraries (but still throws reasonable erros when it fails) we are removing this test.